requirements.txt
icrawler==0.6.6
Pillow>=10.0.0
requests>=2.31.0
